// src/language/generated/ast.ts
export interface ApiDecl extends langium.AstNode {
    readonly $container: App;
    readonly $type: 'ApiDecl';
    authDecls: Array<AuthDecl>;
    method: HttpMethod;
    path: string;
    queries: Array<QueryBlock>;
    returnsDecls: Array<ReturnsDecl>;
}

export const ApiDecl = {
    $type: 'ApiDecl',
    authDecls: 'authDecls',
    method: 'method',
    path: 'path',
    queries: 'queries',
    returnsDecls: 'returnsDecls'
} as const;

export function isApiDecl(item: unknown): item is ApiDecl {
    return reflection.isInstance(item, ApiDecl.$type);
}

export interface App extends langium.AstNode {
    readonly $container: Document;
    readonly $type: 'App';
    apis: Array<ApiDecl>;
    components: Array<ComponentDecl>;
    models: Array<ModelDecl>;
    name: string;
    pages: Array<PageDecl>;
    styles: Array<StyleDecl>;
    vibes: Array<VibeDecl>;
}

export const App = {
    $type: 'App',
    apis: 'apis',
    components: 'components',
    models: 'models',
    name: 'name',
    pages: 'pages',
    styles: 'styles',
    vibes: 'vibes'
} as const;

export function isApp(item: unknown): item is App {
    return reflection.isInstance(item, App.$type);
}

export interface AuthDecl extends langium.AstNode {
    readonly $container: ApiDecl;
    readonly $type: 'AuthDecl';
    level: 'admin' | 'public' | 'user';
}

export const AuthDecl = {
    $type: 'AuthDecl',
    level: 'level'
} as const;

export function isAuthDecl(item: unknown): item is AuthDecl {
    return reflection.isInstance(item, AuthDecl.$type);
}

export interface BoolLit extends langium.AstNode {
    readonly $container: FieldDecl | LoadArg | QueryParam | StyleToken;
    readonly $type: 'BoolLit';
    value: 'false' | 'true';
}

export const BoolLit = {
    $type: 'BoolLit',
    value: 'value'
} as const;
// ...
export interface ChangeDecl extends langium.AstNode {
    readonly $container: Document;
    readonly $type: 'ChangeDecl';
    acceptanceBlocks: Array<AcceptanceBlock>;
    constraintBlocks: Array<ConstraintsBlock>;
    goal: string;
    requirementBlocks: Array<RequirementsBlock>;
    rollbacks: Array<RollbackDecl>;
    scopes: Array<ScopeBlock>;
    styleBlocks: Array<StyleHintsBlock>;
    title: string;
}

export const ChangeDecl = {
    $type: 'ChangeDecl',
    acceptanceBlocks: 'acceptanceBlocks',
    constraintBlocks: 'constraintBlocks',
    goal: 'goal',
    requirementBlocks: 'requirementBlocks',
    rollbacks: 'rollbacks',
    scopes: 'scopes',
    styleBlocks: 'styleBlocks',
    title: 'title'
} as const;

export function isChangeDecl(item: unknown): item is ChangeDecl {
    return reflection.isInstance(item, ChangeDecl.$type);
}

export interface ComponentDecl extends langium.AstNode {
    readonly $container: App;
    readonly $type: 'ComponentDecl';
    elements: Array<UiElement>;
    name: string;
    params: Array<ComponentParam>;
}

export const ComponentDecl = {
    $type: 'ComponentDecl',
    elements: 'elements',
    name: 'name',
    params: 'params'
} as const;

export function isComponentDecl(item: unknown): item is ComponentDecl {
    return reflection.isInstance(item, ComponentDecl.$type);
}

export interface ComponentParam extends langium.AstNode {
    readonly $container: ComponentDecl;
    readonly $type: 'ComponentParam';
    entity: langium.Reference<ModelDecl>;
    name: Name;
}

export const ComponentParam = {
    $type: 'ComponentParam',
    entity: 'entity',
    name: 'name'
} as const;
// ...
export interface Document extends langium.AstNode {
    readonly $type: 'Document';
    apps: Array<App>;
    changes: Array<ChangeDecl>;
}

export const Document = {
    $type: 'Document',
    apps: 'apps',
    changes: 'changes'
} as const;

export function isDocument(item: unknown): item is Document {
    return reflection.isInstance(item, Document.$type);
}

export interface FieldAttr extends langium.AstNode {
    readonly $container: FieldDecl;
    readonly $type: 'FieldAttr';
    arg?: Name;
    name: Name;
}

export const FieldAttr = {
    $type: 'FieldAttr',
    arg: 'arg',
    name: 'name'
} as const;

export function isFieldAttr(item: unknown): item is FieldAttr {
    return reflection.isInstance(item, FieldAttr.$type);
}

export interface FieldDecl extends langium.AstNode {
    readonly $container: ModelDecl;
    readonly $type: 'FieldDecl';
    attrs: Array<FieldAttr>;
    default?: Literal;
    name: Name;
    type: FieldType;
}

export const FieldDecl = {
    $type: 'FieldDecl',
    attrs: 'attrs',
    default: 'default',
    name: 'name',
    type: 'type'
} as const;

export function isFieldDecl(item: unknown): item is FieldDecl {
    return reflection.isInstance(item, FieldDecl.$type);
}

export type FieldType = 'boolean' | 'datetime' | 'decimal' | 'float' | 'int' | 'string' | 'text' | 'uuid';

export function isFieldType(item: unknown): item is FieldType {
    return item === 'uuid' || item === 'string' || item === 'text' || item === 'int' || item === 'float' || item === 'decimal' || item === 'boolean' || item === 'datetime';
}

export interface HeroDecl extends langium.AstNode {
    readonly $container: PageDecl;
    readonly $type: 'HeroDecl';
    cta?: string;
    title?: string;
}

export const HeroDecl = {
    $type: 'HeroDecl',
    cta: 'cta',
    title: 'title'
} as const;

export function isHeroDecl(item: unknown): item is HeroDecl {
    return reflection.isInstance(item, HeroDecl.$type);
}

export type HttpMethod = 'DELETE' | 'GET' | 'PATCH' | 'POST' | 'PUT';

export function isHttpMethod(item: unknown): item is HttpMethod {
    return item === 'GET' || item === 'POST' || item === 'PUT' || item === 'PATCH' || item === 'DELETE';
}

export type Literal = BoolLit | NumberLit | StringLit;

export const Literal = {
    $type: 'Literal'
} as const;

export function isLiteral(item: unknown): item is Literal {
    return reflection.isInstance(item, Literal.$type);
}

export interface LoadArg extends langium.AstNode {
    readonly $container: LoadDecl;
    readonly $type: 'LoadArg';
    name: Name;
    value: Literal;
}

export const LoadArg = {
    $type: 'LoadArg',
    name: 'name',
    value: 'value'
} as const;

export function isLoadArg(item: unknown): item is LoadArg {
    return reflection.isInstance(item, LoadArg.$type);
}

export interface LoadDecl extends langium.AstNode {
    readonly $container: PageDecl;
    readonly $type: 'LoadDecl';
    args: Array<LoadArg>;
    binding: Name;
    method: HttpMethod;
    path: string;
}

export const LoadDecl = {
    $type: 'LoadDecl',
    args: 'args',
    binding: 'binding',
    method: 'method',
    path: 'path'
} as const;

export function isLoadDecl(item: unknown): item is LoadDecl {
    return reflection.isInstance(item, LoadDecl.$type);
}

export interface ModelDecl extends langium.AstNode {
    readonly $container: App;
    readonly $type: 'ModelDecl';
    fields: Array<FieldDecl>;
    name: string;
}

export const ModelDecl = {
    $type: 'ModelDecl',
    fields: 'fields',
    name: 'name'
} as const;

export function isModelDecl(item: unknown): item is ModelDecl {
    return reflection.isInstance(item, ModelDecl.$type);
}

export type Name = 'auth' | 'brand' | 'button' | 'cta' | 'heading' | 'image' | 'price' | 'public' | 'query' | 'text' | 'title' | 'user' | string;

export function isName(item: unknown): item is Name {
    return item === 'price' || item === 'title' || item === 'text' || item === 'image' || item === 'heading' || item === 'button' || item === 'cta' || item === 'brand' || item === 'query' || item === 'auth' || item === 'public' || item === 'user' || (typeof item === 'string' && (/[a-zA-Z_][a-zA-Z0-9_]*/.test(item)));
}

export interface NumberLit extends langium.AstNode {
    readonly $container: FieldDecl | LoadArg | QueryParam | StyleToken;
    readonly $type: 'NumberLit';
    value: number;
}

export const NumberLit = {
    $type: 'NumberLit',
    value: 'value'
} as const;

export function isNumberLit(item: unknown): item is NumberLit {
    return reflection.isInstance(item, NumberLit.$type);
}

export interface PageDecl extends langium.AstNode {
    readonly $container: App;
    readonly $type: 'PageDecl';
    heroes: Array<HeroDecl>;
    loads: Array<LoadDecl>;
    name: string;
    route: string;
    sections: Array<SectionDecl>;
}

export const PageDecl = {
    $type: 'PageDecl',
    heroes: 'heroes',
    loads: 'loads',
    name: 'name',
    route: 'route',
    sections: 'sections'
} as const;
// ...
export interface QueryBlock extends langium.AstNode {
    readonly $container: ApiDecl;
    readonly $type: 'QueryBlock';
    params: Array<QueryParam>;
}

export const QueryBlock = {
    $type: 'QueryBlock',
    params: 'params'
} as const;

export function isQueryBlock(item: unknown): item is QueryBlock {
    return reflection.isInstance(item, QueryBlock.$type);
}

export interface QueryParam extends langium.AstNode {
    readonly $container: QueryBlock;
    readonly $type: 'QueryParam';
    default?: Literal;
    name: Name;
    optional: boolean;
    type: FieldType;
}

export const QueryParam = {
    $type: 'QueryParam',
    default: 'default',
    name: 'name',
    optional: 'optional',
    type: 'type'
} as const;

export function isQueryParam(item: unknown): item is QueryParam {
    return reflection.isInstance(item, QueryParam.$type);
}

export interface RefValue extends langium.AstNode {
    readonly $container: UiElement;
    readonly $type: 'RefValue';
    parts: Array<Name>;
}

export const RefValue = {
    $type: 'RefValue',
    parts: 'parts'
} as const;
// ...
export interface ReturnsDecl extends langium.AstNode {
    readonly $container: ApiDecl;
    readonly $type: 'ReturnsDecl';
    entity: langium.Reference<ModelDecl>;
    many: boolean;
}

export const ReturnsDecl = {
    $type: 'ReturnsDecl',
    entity: 'entity',
    many: 'many'
} as const;
// ...
export interface SectionDecl extends langium.AstNode {
    readonly $container: PageDecl;
    readonly $type: 'SectionDecl';
    binding: Name;
    component: langium.Reference<ComponentDecl>;
    name: string;
}

export const SectionDecl = {
    $type: 'SectionDecl',
    binding: 'binding',
    component: 'component',
    name: 'name'
} as const;
// ...
export interface StringLit extends langium.AstNode {
    readonly $container: FieldDecl | LoadArg | QueryParam | StyleToken;
    readonly $type: 'StringLit';
    value: string;
}

export const StringLit = {
    $type: 'StringLit',
    value: 'value'
} as const;

export function isStringLit(item: unknown): item is StringLit {
    return reflection.isInstance(item, StringLit.$type);
}

export interface StringValue extends langium.AstNode {
    readonly $container: UiElement;
    readonly $type: 'StringValue';
    value: string;
}

export const StringValue = {
    $type: 'StringValue',
    value: 'value'
} as const;

export function isStringValue(item: unknown): item is StringValue {
    return reflection.isInstance(item, StringValue.$type);
}

export interface StyleDecl extends langium.AstNode {
    readonly $container: App;
    readonly $type: 'StyleDecl';
    tokens: Array<StyleToken>;
}

export const StyleDecl = {
    $type: 'StyleDecl',
    tokens: 'tokens'
} as const;
// ...
export interface StyleToken extends langium.AstNode {
    readonly $container: StyleDecl;
    readonly $type: 'StyleToken';
    key: TokenPath;
    value: Literal;
}

export const StyleToken = {
    $type: 'StyleToken',
    key: 'key',
    value: 'value'
} as const;

export function isStyleToken(item: unknown): item is StyleToken {
    return reflection.isInstance(item, StyleToken.$type);
}

export type TokenPath = string;

export function isTokenPath(item: unknown): item is TokenPath {
    return typeof item === 'string';
}

export interface UiElement extends langium.AstNode {
    readonly $container: ComponentDecl;
    readonly $type: 'UiElement';
    kind: 'button' | 'heading' | 'image' | 'price' | 'text';
    value: UiValue;
}

export const UiElement = {
    $type: 'UiElement',
    kind: 'kind',
    value: 'value'
} as const;

export function isUiElement(item: unknown): item is UiElement {
    return reflection.isInstance(item, UiElement.$type);
}

export type UiValue = RefValue | StringValue;

export const UiValue = {
    $type: 'UiValue'
} as const;
// ...
export interface VibeDecl extends langium.AstNode {
    readonly $container: App;
    readonly $type: 'VibeDecl';
    items: Array<VibeItem>;
}

export const VibeDecl = {
    $type: 'VibeDecl',
    items: 'items'
} as const;

export function isVibeDecl(item: unknown): item is VibeDecl {
    return reflection.isInstance(item, VibeDecl.$type);
}

export type VibeItem = A11yItem | BrandItem | PerfItem | SecurityItem | UxItem;

export const VibeItem = {
    $type: 'VibeItem'
} as const;